requirements.txt
Flask==3.0.0
flask-mysqldb==2.0.0
flask-bcrypt==1.0.1
mysqlclient==2.2.0
python-dotenv==1.0.0
Flask-Mail==0.10.0
gunicorn==21.2.0
